Add uploader.py
