+++ src/ansys/fluent/core/session.py
@@ -2,7 +2,6 @@
 import importlib
 import json
 import logging
-import os
 from typing import Any, Callable, Optional  # noqa: F401
 import warnings
 
@@ -30,7 +29,7 @@ from ansys.fluent.core.streaming_services.events_streaming import EventsManager
 from ansys.fluent.core.streaming_services.field_data_streaming import FieldDataStreaming
 from ansys.fluent.core.streaming_services.monitor_streaming import MonitorsManager
 from ansys.fluent.core.streaming_services.transcript_streaming import Transcript
-import ansys.platform.instancemanagement as pypim
+from ansys.fluent.core.utils.uploader import _Uploader
 
 from .rpvars import RPVars
 
@@ -112,7 +111,7 @@ class BaseSession:
         self.error_state = self.fluent_connection.error_state
         self.scheme_eval = self.fluent_connection.scheme_eval
         self.rp_vars = RPVars(self.scheme_eval.string_eval)
-        self._uploader = None
+        self._uploader = _Uploader(self.fluent_connection._remote_instance)
         self._preferences = None
         self.journal = Journal(self.scheme_eval)
 
@@ -272,122 +271,18 @@ class BaseSession:
         logger.debug("session.__exit__() called")
         self.exit()
 
-    def _upload_download_helper(
-        self,
-        is_upload: bool,
-        file_name: Optional[str] = None,
-        remote_file_name: Optional[str] = None,
-        local_file_name: Optional[str] = None,
-    ):
-        """Uploads and downloads a file.
-
-        Parameters
-        ----------
-        is_upload : bool
-            True if we want to upload file, False otherwise.
-        file_name : str
-            file name
-        remote_file_name : str, optional
-            remote file name, by default None
-        local_file_name : str, optional
-            local file name, by default None
-        Raises
-        ------
-        FileNotFoundError
-            If the file does not exist.
-        """
-        if self.file_service:
-            if is_upload:
-                if os.path.isfile(file_name):
-                    expanded_file_path = os.path.expandvars(file_name)
-                    upload_file_name = remote_file_name or os.path.basename(
-                        expanded_file_path
-                    )
-                    self.file_service.upload_file(expanded_file_path, upload_file_name)
-                else:
-                    raise FileNotFoundError(f"{file_name} does not exist.")
-            else:
-                if self.file_service.file_exist(file_name):
-                    self.file_service.download_file(file_name, local_file_name)
-                else:
-                    raise FileNotFoundError("Remote file does not exist.")
-
     def upload(self, file_name: str, remote_file_name: Optional[str] = None):
-        """Uploads a file on the server.
-
-        Parameters
-        ----------
-        file_name : str
-            file name
-        remote_file_name : str, optional
-            remote file name, by default None
-        Raises
-        ------
-        FileNotFoundError
-            If the file does not exist.
-        """
-        self._upload_download_helper(
-            is_upload=True, file_name=file_name, remote_file_name=remote_file_name
-        )
+        """Uploads a file on the server."""
+        return self._uploader.upload(file_name, remote_file_name)
 
     def download(self, file_name: str, local_file_name: Optional[str] = None):
-        """Downloads a file from the server.
-
-        Parameters
-        ----------
-        file_name : str
-            file name
-        local_file_name : str, optional
-            local file path, by default None
-
-        Raises
-        ------
-        FileNotFoundError
-            If the remote file does not exist.
-        """
-        self._upload_download_helper(
-            is_upload=False, file_name=file_name, local_file_name=local_file_name
-        )
+        """Downloads a file from the server."""
+        return self._uploader.download(file_name, local_file_name)
 
     def _pypim_upload(self, file_name: str, on_uploaded: Optional[Any] = None):
-        """Uploads a file if not available on the server.
-
-        Parameters
-        ----------
-        file_name : str
-            File name
-        on_uploaded: Callable[str]
-            Read a file.
-        Raises
-        ------
-        FileNotFoundError
-            If a file does not exist.
-        """
-        if pypim.is_configured():
-            if os.path.isfile(file_name):
-                if not self.file_service.file_exist(os.path.basename(file_name)):
-                    self.upload(file_name)
-            elif self.file_service.file_exist(os.path.basename(file_name)):
-                pass
-            else:
-                raise FileNotFoundError(f"{file_name} does not exist.")
-        if on_uploaded:
-            on_uploaded(os.path.basename(file_name))
+        """If pypim is configured, uploads a file on the server."""
+        return self._uploader._pypim_upload(file_name, on_uploaded)
 
     def _pypim_download(self, file_name: str, on_uploaded: Optional[Any] = None):
-        """Downloads a file from the server.
-
-        Parameters
-        ----------
-        file_name : str
-            File name
-        on_uploaded: Callable[str]
-            Write a file.
-        """
-        if on_uploaded:
-            on_uploaded(os.path.basename(file_name))
-        if pypim.is_configured():
-            if os.path.isfile(file_name):
-                print(f"\nFile already exists. File path:\n{file_name}\n")
-            else:
-                self.download(os.path.basename(file_name), ".")
+        """If pypim is configured, downloads a file from the server."""
+        return self._uploader._pypim_download(file_name, on_uploaded)
